+++ src/apswversion.h
@@ -1,1 +1,1 @@
-#define APSW_VERSION "3.47.1.0"
+#define APSW_VERSION "3.47.2.0"