+++ tools/checksums.py
@@ -11,6 +11,7 @@ sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 import setup
 
 sqlitevers = (
+    "3470200",
     "3470100",
     "3470000",
     "3460100",