+++ doc/changes.rst
@@ -10,8 +10,8 @@ history <https://devguide.python.org/versions/>`__.
 APSW changes by version
 -----------------------
 
-next
-====
+3.47.2.0
+========
 
 Added :func:`apsw.ext.page_usage_to_svg` which shows database usage as
 SVG (`example <_static/samples/chinook.svg>`).  Available as shell